start rewriting benchmarks
