+++ asv_benckmarks/asv.conf.json
@@ -11,7 +11,7 @@
 
     // The URL or local path of the source code repository for the
     // project being benchmarked
-    "repo": ".",
+    "repo": "..",
 
     // The Python project's subdirectory in your repo.  If missing or
     // the empty string, the project is assumed to be located at the root
@@ -47,7 +47,7 @@
 
     // The Pythons you'd like to test against.  If not provided, defaults
     // to the current version of Python used to run `asv`.
-    "pythons": ["3.6","3.7","3.8"],
+    "pythons": ["3.8"],
 
     // The list of conda channel names to be searched for benchmark
     // dependency packages in the specified order
@@ -64,9 +64,9 @@
     // followed by the pip installed packages).
     //
     "matrix": {        // test with and without six installed
-         "numba": ["",null],
-         "numpy":[""],   // emcee is only available for install with pip.
-     },
+        "numba": ["", null],
+    //     "numpy":[""],   // emcee is only available for install with pip.
+    },
 
     // Combinations of libraries/python versions can be excluded/included
     // from the set to test. Each entry is a dictionary containing additional
